feat(catalog-service): add rating to Review entities

+++ catalog_service/src/main/java/feo/health/catalog_service/dto/ReviewDto.java
@@ -11,6 +11,6 @@ public class ReviewDto {
 
     private Date date;
 
-    private Byte rating;
+    private Float rating;
 
 }
